+++ core/Cargo.toml
@@ -18,9 +18,10 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
 
 [features]
-codegen = ["dep:sealed"] # only enables codegen glue
-es = ["dep:derive_more"] # enables event sourcing machinery
-reflect = []             # enables compile-time reflection capabilities
+# Enables event sourcing machinery.
+es = ["dep:derive_more", "dep:sealed"]
+# Enables compile-time reflection capabilities.
+reflect = []
 
 [dependencies]
 derive_more = { version = "0.99", features = ["deref", "deref_mut", "display", "into"], default-features = false, optional = true }
